Clear the open code-scanning alerts (#102)
- The shipped script declared two things it never used: `OWN` in the toast part and
  `setAttr` in the drag part. Both are removed, and the bundle regenerated.
- SednaOverlays returned a dialog's string value through `(TResult)(object)`; a type
  pattern says the same without the cast.
- The catalogue test fixture never disposed the WebApplicationFactory it derived its
  configured one from. The base now owns and disposes it.
- Test assertions compared measured doubles with `==`. Pixel measurements use a
  tolerance, and the hit-test codes, which are whole numbers, compare as integers.
- A redundant cast to the type a variable already had.

+++ src/Sedna.UI.Tests/Browser/ComparisonBarTests.cs
@@ -80,8 +80,8 @@ public class ComparisonBarTests : ScriptTestBase
 
         Assert.Empty(errors);
         Assert.Equal(m[0], m[1], 0.5);
-        Assert.True(m[2] == 1, "The previous value's line is the topmost thing at its own position — over the fill as well.");
-        Assert.True(m[3] == 1 && m[4] == 1, "The line reaches past the track above and below, so it shows when the fill covers the hatch.");
+        Assert.True((int)m[2] == 1, "The previous value's line is the topmost thing at its own position — over the fill as well.");
+        Assert.True((int)m[3] == 1 && (int)m[4] == 1, "The line reaches past the track above and below, so it shows when the fill covers the hatch.");
         Assert.True(m[5] > m[6], $"The line is {m[5]}px tall on a {m[6]}px track.");
     }
